moved SimpleClient to another py file
